split gluster and ceph tests and api
