+++ usmqe_tests/api/tendrl/README.rst
@@ -1,0 +1,3 @@
+There are 2 types of tests here:
+1) Gluster ones in ```gluster``` directory and
+2) Ceph ones in ```ceph``` directory.
